got-ocr-server/requirements.txt
torch==2.0.1
torchvision==0.15.2
transformers==4.37.2
tiktoken==0.6.0
verovio==4.3.1
accelerate==0.28.0
flask==3.0.0
flask-cors==4.0.0
pillow==10.2.0
numpy==1.24.3